Uploaded from main server pc
This is the excel file where I've created a rate query file along with a attached databased. This database is creating when the query is send via mail. Moreover, the file generate a random number, so that, we can easily track & trace the mail & this number is unique to the database also. This file have a dashboard for better visualization of data for decision making. The dashboard is automatically updated from the database data.